fix warning

+++ src/platform/platform.cpp
@@ -43,8 +43,7 @@ void GLRenderCtx::cleanup() {
 GLRenderCtx::GLRenderCtx(ImVec2 const& size) : m_size(size) {}
 
 ImTextureID GLRenderCtx::texture() const {
-    // TODO: this wont work on 64 bit
-    return reinterpret_cast<ImTextureID>(m_texture);
+    return reinterpret_cast<ImTextureID>(static_cast<uintptr_t>(m_texture));
 }
 
 ImVec2 GLRenderCtx::size() const {
